Refactor CSS modules and functions
