+++ src/Components/PaperEdits/PaperEdit/ProgramScript/PreviewCanvas/Filler.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './VideoContextProgressBar.module.css';
+import styles from './Filler.module.css';
 
 const Filler = (props) => {
   return <div className={ styles.filler } style={ { width: `${ props.percentage }%` } }></div>;
