+++ src/Components/PaperEdits/PaperEdit/ProgramScript/PreviewCanvas/Filler.module.css
@@ -1,0 +1,6 @@
+
+.filler {
+    background: red;
+    height: 100%;
+    border-radius: inherit;
+  }
