+++ src/Components/PaperEdits/PaperEdit/ProgramScript/PreviewCanvas/ProgressBar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Filler from './Filler';
-import styles from './VideoContextProgressBar.module.css';
+import styles from './ProgressBar.module.css';
 
 const ProgressBar = (props) => {
