+++ src/Components/PaperEdits/PaperEdit/ProgramScript/PreviewCanvas/ProgressBar.module.css
@@ -1,0 +1,8 @@
+
+.progressBar {
+    position: relative;
+    height: 12px;
+    width: 100%;
+    opacity: 1;
+  }
+
